requirements.txt
streamlit>=1.28.0
msal>=1.24.0
python-dotenv>=1.0.0
streamlit-cookie-controller>=0.0.2